share the ringing activity behavior and layout
Extract lock-screen visibility, display wake handling, face-down actions, and volume key behavior so other ringing events can reuse them. Keep alarm-specific dismissal and snooze behavior in the alarm activity.

Share the dark layout and landscape arrangement while accepting the displayed time, label, and ringing icon from each activity.
