+++ app/src/main/java/com/bnyro/clock/presentation/screens/alarm/AlarmAlertScreen.kt
@@ -1,27 +1,13 @@
 package com.bnyro.clock.presentation.screens.alarm
 
 import android.content.res.Configuration
-import android.content.res.Configuration.ORIENTATION_PORTRAIT
-import androidx.compose.animation.core.Ease
-import androidx.compose.animation.core.EaseInOutBack
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Add
@@ -32,32 +18,21 @@ import androidx.compose.material3.FilledTonalButton
 import androidx.compose.material3.FilledTonalIconButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.bnyro.clock.R
-import com.bnyro.clock.presentation.screens.settings.model.SettingsModel
-import com.bnyro.clock.ui.theme.ClockYouTheme
-import com.bnyro.clock.util.ThemeUtil
-import com.bnyro.clock.util.TimeHelper
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.isActive
+import com.bnyro.clock.presentation.screens.ringing.RingingAlert
+import com.bnyro.clock.presentation.screens.ringing.RingingTitle
 
 @Composable
 fun AlarmAlertScreen(
@@ -67,52 +42,8 @@ fun AlarmAlertScreen(
     snoozeEnabled: Boolean,
     snoozeTime: Int
 ) {
-    val settingsModel: SettingsModel = viewModel()
-    ClockYouTheme(
-        darkTheme = true,
-        customColorScheme = ThemeUtil.getSchemeFromSeed(
-            settingsModel.customColor,
-            true
-        )
-    ) {
-        val orientation = LocalConfiguration.current.orientation
-        Surface(
-            modifier = Modifier.fillMaxSize(),
-            color = MaterialTheme.colorScheme.background
-        ) {
-            if (orientation == ORIENTATION_PORTRAIT) {
-                Column(
-                    modifier = Modifier.fillMaxSize(),
-                    verticalArrangement = Arrangement.SpaceEvenly,
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    AlarmAnimation()
-                    AlarmControls(label, snoozeTime, snoozeEnabled, onSnooze, onDismiss)
-                }
-            } else {
-                Row {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .weight(2f),
-                        verticalArrangement = Arrangement.Center,
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        AlarmAnimation()
-                    }
-
-                    Column(
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .weight(3f),
-                        verticalArrangement = Arrangement.SpaceEvenly,
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        AlarmControls(label, snoozeTime, snoozeEnabled, onSnooze, onDismiss)
-                    }
-                }
-            }
-        }
+    RingingAlert(painterResource(id = R.drawable.ic_alarm)) {
+        AlarmControls(label, snoozeTime, snoozeEnabled, onSnooze, onDismiss)
     }
 }
 
@@ -124,23 +55,7 @@ private fun AlarmControls(
     onSnooze: (minutes: Int) -> Unit,
     onDismiss: () -> Unit
 ) {
-    val context = LocalContext.current
-    val time by produceState(
-        initialValue = TimeHelper.formatTime(context, TimeHelper.getTimeByZone()),
-        producer = {
-            while (isActive) {
-                value = TimeHelper.formatTime(context, TimeHelper.getTimeByZone())
-                delay(1000)
-            }
-        }
-    )
-    Text(
-        text = time,
-        style = MaterialTheme.typography.displayMedium
-    )
-    label?.let {
-        Text(text = it, style = MaterialTheme.typography.headlineMedium)
-    }
+    RingingTitle(label)
     Column(
         Modifier.fillMaxWidth(),
         horizontalAlignment = Alignment.CenterHorizontally
@@ -199,42 +114,6 @@ private fun AlarmControls(
                 }
             }
         }
-    }
-}
-
-@Composable
-private fun AlarmAnimation() {
-    val infiniteTransition = rememberInfiniteTransition(label = "")
-    val rotation by infiniteTransition.animateFloat(
-        initialValue = -10F,
-        targetValue = 10F,
-        animationSpec = infiniteRepeatable(
-            animation = tween(400, easing = EaseInOutBack),
-            repeatMode = RepeatMode.Reverse
-        ),
-        label = ""
-    )
-    val offset by infiniteTransition.animateFloat(
-        initialValue = 10F,
-        targetValue = -10F,
-        animationSpec = infiniteRepeatable(
-            animation = tween(200, easing = Ease),
-            repeatMode = RepeatMode.Reverse
-        ),
-        label = ""
-    )
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = Modifier
-            .offset(y = offset.dp)
-            .rotate(rotation)
-    ) {
-        Image(
-            modifier = Modifier.size(250.dp),
-            painter = painterResource(id = R.drawable.ic_alarm),
-            contentDescription = null,
-            colorFilter = ColorFilter.tint(MaterialTheme.colorScheme.primary)
-        )
     }
 }
